added basic collection tests for references
