+++ test/references/index.js
@@ -1,0 +1,3 @@
+'use strict'
+require('./basic')
+require('./collection')
