+++ test/test.js
@@ -11,6 +11,7 @@ module.exports = function (t, state, subs) {
     state,
     subs,
     function (state, type, stamp, subs, tree, sType) {
+      console.log('FIRE SUBS', state.path().join('/'), type, stamp)
       updates.push({
         path: state.path().join('/'),
         type: type,
